Add files via upload
Fresh fruits and vegetables are in greater demand as a result of rising living standards. Due to their perishable nature, it is extremely difficult to preserve the nutritional value and freshness of agricultural products after harvest. Inadequate processing, a lack of suitable refrigerated equipment, and inadequate cold chain logistics facilities have led to significant losses in fruits and vegetables. In the world, food is lost at a rate of 40–50% per year for fruits and vegetables. Low temperature slows bacterial proliferation, respiratory intensity, and metabolism, which delays spoiling (Lagarda-Leyva et al, 2023). This model determines the appropriate number of forklifts and specialist handling machines needed to meet service standards when dealing with food shipments in order to avoid the losses